refactor: rename & move types files to 'types' folders
