+++ src/app/core/components/social-buttons/mockSocialData.ts
@@ -1,4 +1,4 @@
-import { User } from '@core/models/user.model';
+import { User } from '@core/types/user';
 
 export const mockSocialData: User = {
   id: 42,
