+++ src/app/core/services/airports.service.ts
@@ -2,7 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable, map } from 'rxjs';
 import { environment } from 'src/environments/environment';
-import { Airport } from '@core/interfaces/airport';
+import { Airport } from '@booking/types/airport';
 
 @Injectable({
   providedIn: 'root',
