+++ src/app/core/types/social-data.ts
