+++ src/app/flight-search/pages/flight-search-page/flight-search-page.component.ts
@@ -4,8 +4,8 @@ import { Observable, Subject, Subscription, debounceTime, distinctUntilChanged, 
 import { TuiDay } from '@taiga-ui/cdk';
 import plur from 'plur';
 import { AirportsService } from '@core/services/airports.service';
-import { Airport } from '@core/interfaces/airport';
-import { PassengerCategory, Passengers } from '@shared/interfaces/passengers';
+import { Airport } from '@booking/types/airport';
+import { PassengerCategory, Passengers } from '@shared/types/passengers';
 import { FlightType } from '@shared/types/flight-type';
 import { adultValidator } from '@flight-search/validators/adultValidator';
 import { FLIGHT_SEARCH_MINIMUM_QUERY_LENGTH, FLIGHT_SEARCH_DEBOUNCE_TIME } from '@flight-search/constants';
