--- src/app/shared/interfaces/passengers.ts
@@ -1,7 +1,0 @@
-export interface Passengers {
-  adults: number;
-  children: number;
-  infants: number;
-}
-
-export type PassengerCategory = keyof Passengers;
