+++ src/app/user/pages/shopping-cart-page/shopping-cart-page.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
-import { Booking } from '@shared/models/booking.model';
+import { Booking } from '@shared/types/booking';
 import { mockBookingsData } from './mockBookingsData';
 
 @Component({
